pyproject.toml
[tool.poetry]
name = "nina"
version = "0.0.1"
description = ""
authors = []
license = ""

[tool.poetry.dependencies]
python = "^3.6.7"
awscli = "=1.16.166"
boto3 = "=1.9.156"
python-decouple = "=3.1"
python-levenshtein = "=0.12.0"
pandas = "=0.24.2"
joblib = "=0.13.2"
records = "=0.5.3"
s3fs = "=0.4.0"
pyarrow = "=0.14.1"
PyAthena = "=1.8.0"
autopep8 = "=1.4.4"
ipdb = "=0.12"
jupyter = "=1.0.0"
pep8 = "=1.7.1"
pylint = "=2.3.1"
pytest = "=4.5.0"
flake8 = "=3.7.9"
rope = "=0.14.0"
requests = "=2.22.0"
tqdm = "=4.36.1"
pandas-profiling = "=2.3.0"
datacompy = "=0.6.0"